@@ -5,8 +5,11 @@ import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.os.Build;
 import android.os.Bundle;
+
 import androidx.appcompat.app.AppCompatActivity;
+
 import android.view.View;
+
 import com.qiniu.pili.droid.shortvideo.PLAuthenticationResultCallback;
 import com.qiniu.pili.droid.shortvideo.PLShortVideoEnv;
 import com.qiniu.pili.droid.shortvideo.demo.BuildConfig;
@@ -30,17 +33,17 @@ public class MainActivity extends AppCompatActivity {
         itemVersionInfo.setValue("" + getVersionDescription());
 
         ItemView itemCompileInfo = (ItemView) findViewById(R.id.item_compile_info);
-        itemCompileInfo.setValue(""+ getBuildTimeDescription());
+        itemCompileInfo.setValue("" + getBuildTimeDescription());
 
         PLShortVideoEnv.checkAuthentication(getApplicationContext(), new PLAuthenticationResultCallback() {
             @Override
             public void onAuthorizationResult(int result) {
                 if (result == PLAuthenticationResultCallback.UnCheck) {
-                    ToastUtils.s(MainActivity.this, "UnCheck");
+                    ToastUtils.showShortToast(MainActivity.this, "UnCheck");
                 } else if (result == PLAuthenticationResultCallback.UnAuthorized) {
-                    ToastUtils.s(MainActivity.this, "UnAuthorized");
+                    ToastUtils.showShortToast(MainActivity.this, "UnAuthorized");
                 } else {
-                    ToastUtils.s(MainActivity.this, "Authorized");
+                    ToastUtils.showShortToast(MainActivity.this, "Authorized");
                 }
             }
         });
@@ -55,7 +58,7 @@ public class MainActivity extends AppCompatActivity {
         PermissionChecker checker = new PermissionChecker(this);
         boolean isPermissionOK = Build.VERSION.SDK_INT < Build.VERSION_CODES.M || checker.checkPermission();
         if (!isPermissionOK) {
-            ToastUtils.s(this, "Some permissions is not approved !!!");
+            ToastUtils.showShortToast(this, "Some permissions is not approved !!!");
         }
         return isPermissionOK;
     }
@@ -72,8 +75,8 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
-    public void onClickImportAndRecord(View v){
-        if (isPermissionOK()){
+    public void onClickImportAndRecord(View v) {
+        if (isPermissionOK()) {
             Intent intent = new Intent(this, ImportAndEditActivity.class);
             startActivity(intent);
         }
@@ -138,7 +141,7 @@ public class MainActivity extends AppCompatActivity {
         intent.putExtra(VideoRecordActivity.PREVIEW_SIZE_RATIO, ConfigActivity.PREVIEW_SIZE_RATIO_POS);
         intent.putExtra(VideoRecordActivity.PREVIEW_SIZE_LEVEL, ConfigActivity.PREVIEW_SIZE_LEVEL_POS);
         intent.putExtra(VideoRecordActivity.ENCODING_MODE, ConfigActivity.ENCODING_MODE_LEVEL_POS);
-        intent.putExtra(VideoRecordActivity.ENCODING_SIZE_LEVEL,ConfigActivity.ENCODING_SIZE_LEVEL_POS);
+        intent.putExtra(VideoRecordActivity.ENCODING_SIZE_LEVEL, ConfigActivity.ENCODING_SIZE_LEVEL_POS);
         intent.putExtra(VideoRecordActivity.ENCODING_BITRATE_LEVEL, ConfigActivity.ENCODING_BITRATE_LEVEL_POS);
         intent.putExtra(VideoRecordActivity.AUDIO_CHANNEL_NUM, ConfigActivity.AUDIO_CHANNEL_NUM_POS);
         startActivity(intent);
