@@ -2,7 +2,9 @@ package com.qiniu.pili.droid.shortvideo.demo.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
+
 import androidx.appcompat.app.AppCompatActivity;
+
 import android.text.TextUtils;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
@@ -25,24 +27,13 @@ public class WebViewActivity extends AppCompatActivity {
             Toast.makeText(this, "地址为空", Toast.LENGTH_SHORT).show();
         } else {
             WebView webView = (WebView) findViewById(R.id.web);
-            webView.setWebViewClient(new WebViewClient() {
-                @Override
-                public boolean shouldOverrideUrlLoading(WebView view, String url) {
-                    return false;// 返回false
-                }
-            });
+            webView.setWebViewClient(new WebViewClient());
 
             WebSettings webSettings = webView.getSettings();
             // 让WebView能够执行javaScript
             webSettings.setJavaScriptEnabled(true);
             // 让JavaScript可以自动打开windows
             webSettings.setJavaScriptCanOpenWindowsAutomatically(true);
-//        // 设置缓存
-//        webSettings.setAppCacheEnabled(true);
-//        // 设置缓存模式,一共有四种模式
-//        webSettings.setCacheMode(WebSettings.LOAD_CACHE_ELSE_NETWORK);
-//        // 设置缓存路径
-//        webSettings.setAppCachePath("/storage/emulated/0/Android/data/com.easyar.buddha/files");
             // 支持缩放(适配到当前屏幕)
             webSettings.setSupportZoom(true);
             // 将图片调整到合适的大小
