@@ -2,9 +2,12 @@ package com.qiniu.pili.droid.shortvideo.demo.view;
 
 import android.content.Context;
 import android.os.Bundle;
+
 import androidx.annotation.NonNull;
+
 import com.google.android.material.bottomsheet.BottomSheetBehavior;
 import com.google.android.material.bottomsheet.BottomSheetDialog;
+
 import android.view.View;
 import android.widget.SeekBar;
 
@@ -31,7 +34,7 @@ public class AudioMixSettingDialog extends BottomSheetDialog {
         void onPositionSelected(long position);
     }
 
-    public AudioMixSettingDialog(Context context){
+    public AudioMixSettingDialog(Context context) {
         super(context);
     }
 
