@@ -26,7 +26,7 @@ import com.pili.pldroid.player.IMediaController;
 
 import java.util.Locale;
 
-public class MediaController extends FrameLayout implements IMediaController{
+public class MediaController extends FrameLayout implements IMediaController {
     private static final String TAG = "PLMediaController";
     private IMediaController.MediaPlayerControl mPlayer;
     private Context mContext;
@@ -54,17 +54,17 @@ public class MediaController extends FrameLayout implements IMediaController{
 
     private boolean mUseFastForward;
 
-    private static final int IC_MEDIA_PAUSE_ID = Resources.getSystem().getIdentifier("ic_media_pause","drawable", "android");
-    private static final int IC_MEDIA_PLAY_ID = Resources.getSystem().getIdentifier("ic_media_play","drawable", "android");
+    private static final int IC_MEDIA_PAUSE_ID = Resources.getSystem().getIdentifier("ic_media_pause", "drawable", "android");
+    private static final int IC_MEDIA_PLAY_ID = Resources.getSystem().getIdentifier("ic_media_play", "drawable", "android");
     private static final int MEDIA_CONTROLLER_ID = Resources.getSystem().getIdentifier("media_controller", "layout", "android");
-    private static final int PRV_BUTTON_ID = Resources.getSystem().getIdentifier("prev","id", "android");
-    private static final int FFWD_BUTTON_ID = Resources.getSystem().getIdentifier("ffwd","id", "android");
-    private static final int NEXT_BUTTON_ID = Resources.getSystem().getIdentifier("next","id", "android");
-    private static final int REW_BUTTON_ID = Resources.getSystem().getIdentifier("rew","id", "android");
-    private static final int PAUSE_BUTTON_ID = Resources.getSystem().getIdentifier("pause","id", "android");
-    private static final int MEDIACONTROLLER_PROGRESS_ID = Resources.getSystem().getIdentifier("mediacontroller_progress","id", "android");
-    private static final int END_TIME_ID = Resources.getSystem().getIdentifier("time","id", "android");
-    private static final int CURRENT_TIME_ID = Resources.getSystem().getIdentifier("time_current","id", "android");
+    private static final int PRV_BUTTON_ID = Resources.getSystem().getIdentifier("prev", "id", "android");
+    private static final int FFWD_BUTTON_ID = Resources.getSystem().getIdentifier("ffwd", "id", "android");
+    private static final int NEXT_BUTTON_ID = Resources.getSystem().getIdentifier("next", "id", "android");
+    private static final int REW_BUTTON_ID = Resources.getSystem().getIdentifier("rew", "id", "android");
+    private static final int PAUSE_BUTTON_ID = Resources.getSystem().getIdentifier("pause", "id", "android");
+    private static final int MEDIACONTROLLER_PROGRESS_ID = Resources.getSystem().getIdentifier("mediacontroller_progress", "id", "android");
+    private static final int END_TIME_ID = Resources.getSystem().getIdentifier("time", "id", "android");
+    private static final int CURRENT_TIME_ID = Resources.getSystem().getIdentifier("time_current", "id", "android");
 
     private AudioManager mAM;
     private Runnable mLastSeekBarRunnable;
@@ -73,7 +73,9 @@ public class MediaController extends FrameLayout implements IMediaController{
 
     public interface OnClickSpeedAdjustListener {
         void onClickNormal();
+
         void onClickFaster();
+
         void onClickSlower();
     }
 
@@ -86,8 +88,9 @@ public class MediaController extends FrameLayout implements IMediaController{
 
     public MediaController(Context context) {
         super(context);
-        if (!mFromXml && initController(context))
+        if (!mFromXml && initController(context)) {
             initFloatingWindow();
+        }
     }
 
     public MediaController(Context context, boolean useFastForward, boolean disableProgressBar) {
@@ -114,8 +117,9 @@ public class MediaController extends FrameLayout implements IMediaController{
 
     @Override
     public void onFinishInflate() {
-        if (mRoot != null)
+        if (mRoot != null) {
             initControllerView(mRoot);
+        }
         super.onFinishInflate();
     }
 
@@ -188,8 +192,7 @@ public class MediaController extends FrameLayout implements IMediaController{
     /**
      * Control the action when the seekbar dragged by user
      *
-     * @param seekWhenDragging
-     * True the media will seek periodically
+     * @param seekWhenDragging True the media will seek periodically
      */
     public void setInstantSeeking(boolean seekWhenDragging) {
         mInstantSeeking = seekWhenDragging;
@@ -197,8 +200,9 @@ public class MediaController extends FrameLayout implements IMediaController{
 
     private void disableUnsupportedButtons() {
         try {
-            if (mPauseButton != null && !mPlayer.canPause())
+            if (mPauseButton != null && !mPlayer.canPause()) {
                 mPauseButton.setEnabled(false);
+            }
         } catch (IncompatibleClassChangeError ex) {
         }
     }
@@ -213,11 +217,9 @@ public class MediaController extends FrameLayout implements IMediaController{
      * the next time the controller is shown.
      * </p>
      *
-     * @param animationStyle
-     * animation style to use when the controller appears and disappears.
-     * Set to -1 for the default animation, 0 for no animation,
-     * or a resource identifier for an explicit animation.
-     *
+     * @param animationStyle animation style to use when the controller appears and disappears.
+     *                       Set to -1 for the default animation, 0 for no animation,
+     *                       or a resource identifier for an explicit animation.
      */
     public void setAnimationStyle(int animationStyle) {
         mAnimStyle = animationStyle;
@@ -243,8 +245,7 @@ public class MediaController extends FrameLayout implements IMediaController{
         mHiddenListener = l;
     }
 
-    @SuppressLint("HandlerLeak")
-    private Handler mHandler = new Handler() {
+    @SuppressLint("HandlerLeak") private Handler mHandler = new Handler() {
         @Override
         public void handleMessage(Message msg) {
             long pos;
@@ -265,8 +266,9 @@ public class MediaController extends FrameLayout implements IMediaController{
     };
 
     private long setProgress() {
-        if (mPlayer == null || mDragging)
+        if (mPlayer == null || mDragging) {
             return 0;
+        }
 
         long position = mPlayer.getCurrentPosition();
         long duration = mPlayer.getDuration();
@@ -281,10 +283,12 @@ public class MediaController extends FrameLayout implements IMediaController{
 
         mDuration = duration;
 
-        if (mEndTime != null)
+        if (mEndTime != null) {
             mEndTime.setText(generateTime(mDuration));
-        if (mCurrentTime != null)
+        }
+        if (mCurrentTime != null) {
             mCurrentTime.setText(generateTime(position));
+        }
 
         return position;
     }
@@ -297,11 +301,9 @@ public class MediaController extends FrameLayout implements IMediaController{
         int hours = totalSeconds / 3600;
 
         if (hours > 0) {
-            return String.format(Locale.US, "%02d:%02d:%02d", hours, minutes,
-                    seconds).toString();
+            return String.format(Locale.US, "%02d:%02d:%02d", hours, minutes, seconds);
         } else {
-            return String.format(Locale.US, "%02d:%02d", minutes, seconds)
-                    .toString();
+            return String.format(Locale.US, "%02d:%02d", minutes, seconds);
         }
     }
 
@@ -325,8 +327,9 @@ public class MediaController extends FrameLayout implements IMediaController{
                 || keyCode == KeyEvent.KEYCODE_MEDIA_PLAY_PAUSE || keyCode == KeyEvent.KEYCODE_SPACE)) {
             doPauseResume();
             show(sDefaultTimeout);
-            if (mPauseButton != null)
+            if (mPauseButton != null) {
                 mPauseButton.requestFocus();
+            }
             return true;
         } else if (keyCode == KeyEvent.KEYCODE_MEDIA_STOP) {
             if (mPlayer.isPlaying()) {
@@ -334,8 +337,7 @@ public class MediaController extends FrameLayout implements IMediaController{
                 updatePausePlay();
             }
             return true;
-        } else if (keyCode == KeyEvent.KEYCODE_BACK
-                || keyCode == KeyEvent.KEYCODE_MENU) {
+        } else if (keyCode == KeyEvent.KEYCODE_BACK || keyCode == KeyEvent.KEYCODE_MENU) {
             hide();
             return true;
         } else {
@@ -345,6 +347,7 @@ public class MediaController extends FrameLayout implements IMediaController{
     }
 
     private View.OnClickListener mPauseListener = new View.OnClickListener() {
+        @Override
         public void onClick(View v) {
             if (mOnClickSpeedAdjustListener != null) {
                 mOnClickSpeedAdjustListener.onClickNormal();
@@ -355,36 +358,43 @@ public class MediaController extends FrameLayout implements IMediaController{
     };
 
     private void updatePausePlay() {
-        if (mRoot == null || mPauseButton == null)
+        if (mRoot == null || mPauseButton == null) {
             return;
+        }
 
-        if (mPlayer.isPlaying())
+        if (mPlayer.isPlaying()) {
             mPauseButton.setImageResource(IC_MEDIA_PAUSE_ID);
-        else
+        } else {
             mPauseButton.setImageResource(IC_MEDIA_PLAY_ID);
+        }
     }
 
     private void doPauseResume() {
-        if (mPlayer.isPlaying())
+        if (mPlayer.isPlaying()) {
             mPlayer.pause();
-        else
+        } else {
             mPlayer.start();
+        }
         updatePausePlay();
     }
 
     private SeekBar.OnSeekBarChangeListener mSeekListener = new SeekBar.OnSeekBarChangeListener() {
 
+        @Override
         public void onStartTrackingTouch(SeekBar bar) {
             mDragging = true;
             show(3600000);
             mHandler.removeMessages(SHOW_PROGRESS);
-            if (mInstantSeeking)
+            if (mInstantSeeking) {
                 mAM.setStreamMute(AudioManager.STREAM_MUSIC, true);
+            }
         }
 
+        @Override
         public void onProgressChanged(SeekBar bar, int progress, boolean fromuser) {
-            if (!fromuser)
+            if (!fromuser) {
                 return;
+            }
 
             final long newposition = (long) (mDuration * progress) / 1000;
             String time = generateTime(newposition);
@@ -398,13 +408,16 @@ public class MediaController extends FrameLayout implements IMediaController{
                 };
                 mHandler.postDelayed(mLastSeekBarRunnable, SEEK_TO_POST_DELAY_MILLIS);
             }
-            if (mCurrentTime != null)
+            if (mCurrentTime != null) {
                 mCurrentTime.setText(time);
+            }
         }
 
+        @Override
         public void onStopTrackingTouch(SeekBar bar) {
-            if (!mInstantSeeking)
+            if (!mInstantSeeking) {
                 mPlayer.seekTo(mDuration * bar.getProgress() / 1000);
+            }
 
             show(sDefaultTimeout);
             mHandler.removeMessages(SHOW_PROGRESS);
@@ -415,6 +428,7 @@ public class MediaController extends FrameLayout implements IMediaController{
     };
 
     private View.OnClickListener mRewListener = new View.OnClickListener() {
+        @Override
         public void onClick(View v) {
             if (mOnClickSpeedAdjustListener != null) {
                 mOnClickSpeedAdjustListener.onClickSlower();
@@ -424,6 +438,7 @@ public class MediaController extends FrameLayout implements IMediaController{
     };
 
     private View.OnClickListener mFfwdListener = new View.OnClickListener() {
+        @Override
         public void onClick(View v) {
             if (mOnClickSpeedAdjustListener != null) {
                 mOnClickSpeedAdjustListener.onClickFaster();
@@ -434,12 +449,11 @@ public class MediaController extends FrameLayout implements IMediaController{
 
     /**
      * Set the view that acts as the anchor for the control view.
-     *
+     * <p>
      * - This can for example be a VideoView, or your Activity's main view.
      * - AudioPlayer has no anchor view, so the view parameter will be null.
      *
-     * @param view
-     * The view to which to anchor the controller when it is visible.
+     * @param view The view to which to anchor the controller when it is visible.
      */
     @Override
     public void setAnchorView(View view) {
@@ -472,8 +486,7 @@ public class MediaController extends FrameLayout implements IMediaController{
      * Show the controller on screen. It will go away automatically after
      * 'timeout' milliseconds of inactivity.
      *
-     * @param timeout
-     * The timeout in milliseconds. Use 0 to show the controller until hide() is called.
+     * @param timeout The timeout in milliseconds. Use 0 to show the controller until hide() is called.
      */
     @Override
     public void show(int timeout) {
@@ -483,8 +496,9 @@ public class MediaController extends FrameLayout implements IMediaController{
                     mAnchor.setSystemUiVisibility(View.SYSTEM_UI_FLAG_VISIBLE);
                 }
             }
-            if (mPauseButton != null)
+            if (mPauseButton != null) {
                 mPauseButton.requestFocus();
+            }
             disableUnsupportedButtons();
 
             if (mFromXml) {
@@ -499,29 +513,27 @@ public class MediaController extends FrameLayout implements IMediaController{
                             + mAnchor.getHeight());
 
                     mWindow.setAnimationStyle(mAnimStyle);
-                    mWindow.showAtLocation(mAnchor, Gravity.BOTTOM,
-                            anchorRect.left, 0);
+                    mWindow.showAtLocation(mAnchor, Gravity.BOTTOM, anchorRect.left, 0);
                 } else {
                     Rect anchorRect = new Rect(location[0], location[1],
                             location[0] + mRoot.getWidth(), location[1]
                             + mRoot.getHeight());
 
                     mWindow.setAnimationStyle(mAnimStyle);
-                    mWindow.showAtLocation(mRoot, Gravity.BOTTOM,
-                            anchorRect.left, 0);
+                    mWindow.showAtLocation(mRoot, Gravity.BOTTOM, anchorRect.left, 0);
                 }
             }
             mShowing = true;
-            if (mShownListener != null)
+            if (mShownListener != null) {
                 mShownListener.onShown();
+            }
         }
         updatePausePlay();
         mHandler.sendEmptyMessage(SHOW_PROGRESS);
 
         if (timeout != 0) {
             mHandler.removeMessages(FADE_OUT);
-            mHandler.sendMessageDelayed(mHandler.obtainMessage(FADE_OUT),
-                    timeout);
+            mHandler.sendMessageDelayed(mHandler.obtainMessage(FADE_OUT), timeout);
         }
     }
 
@@ -533,23 +545,20 @@ public class MediaController extends FrameLayout implements IMediaController{
     @Override
     public void hide() {
         if (mShowing) {
-            if (mAnchor != null) {
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.ICE_CREAM_SANDWICH) {
-                    //mAnchor.setSystemUiVisibility(View.SYSTEM_UI_FLAG_HIDE_NAVIGATION);
-                }
-            }
             try {
                 mHandler.removeMessages(SHOW_PROGRESS);
-                if (mFromXml)
+                if (mFromXml) {
                     setVisibility(View.GONE);
-                else
+                } else {
                     mWindow.dismiss();
+                }
             } catch (IllegalArgumentException ex) {
                 Log.d(TAG, "MediaController already removed");
             }
             mShowing = false;
-            if (mHiddenListener != null)
+            if (mHiddenListener != null) {
                 mHiddenListener.onHidden();
+            }
         }
     }
 
@@ -564,8 +573,9 @@ public class MediaController extends FrameLayout implements IMediaController{
         if (mRewButton != null) {
             mRewButton.setEnabled(enabled);
         }
-        if (mProgress != null && !mDisableProgress)
+        if (mProgress != null && !mDisableProgress) {
             mProgress.setEnabled(enabled);
+        }
         disableUnsupportedButtons();
         super.setEnabled(enabled);
     }
